backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic[email]==2.7.1
python-dotenv==1.0.1
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
bcrypt==4.0.1
pyotp==2.9.0
httpx==0.27.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.128.0
